Move legend style types to root types.ts, import from there in components
