+++ projects/js-packages/charts/src/components/legend/types.ts
@@ -1,44 +1,19 @@
 import { LegendOrdinal } from '@visx/legend';
+import type {
+	LegendItemStyles,
+	LegendLabelStyles,
+	LegendPosition,
+	LegendShapeStyles,
+} from '../../types';
 import type { GlyphProps, LineStyles } from '@visx/xychart';
 import type { ComponentProps, CSSProperties, ReactNode } from 'react';
+
+export type { LegendItemStyles, LegendLabelStyles, LegendPosition, LegendShapeStyles };
 
 type VisxLegendProps = Pick<
 	ComponentProps< typeof LegendOrdinal >,
 	'className' | 'shape' | 'fill' | 'size' | 'labelFormat' | 'labelTransform'
 >;
-
-export type LegendItemStyles = {
-	/** Margin around each legend item. */
-	margin?: CSSProperties[ 'margin' ];
-	/** Flex direction for items within each legend entry. */
-	flexDirection?: 'row' | 'row-reverse' | 'column' | 'column-reverse';
-};
-
-export type LegendLabelStyles = Pick< CSSProperties, 'justifyContent' | 'flex' | 'margin' > & {
-	/**
-	 * Maximum width for legend label text as a CSS value (e.g. '200px', '50%', '10rem').
-	 * When set, text overflow behavior is controlled by textOverflow.
-	 */
-	maxWidth?: string;
-	/**
-	 * Controls how text behaves when it exceeds maxWidth.
-	 * - 'ellipsis': Truncate with ellipsis (ideal for widgets/small devices)
-	 * - 'wrap': Wrap text to multiple lines (default, ideal for larger displays)
-	 */
-	textOverflow?: 'ellipsis' | 'wrap';
-};
-
-export type LegendShapeStyles = {
-	/** Width of the legend shape in pixels. */
-	width?: number;
-	/** Height of the legend shape in pixels. */
-	height?: number;
-	/** Margin around the legend shape. */
-	margin?: CSSProperties[ 'margin' ];
-};
-
-/** Position of the legend relative to chart content. */
-export type LegendPosition = 'top' | 'bottom';
 
 export type BaseLegendProps = VisxLegendProps & {
 	items: BaseLegendItem[];
